- AdvertisementOperationJPAImpl implementation started int id-s converted to long

+++ VirtualShelter/src/main/java/hu/bme/aut/vshelter/dal/AdvertisementFacade.java
@@ -14,6 +14,6 @@ public interface AdvertisementFacade {
 	
 	void edit(Advertisement advertisement);
 	
-	void deleteAdvertisementById(int advertisementId);
+	void deleteAdvertisementById(long advertisementId);
 
 }
